+++ internal/filestorage/diskFileStorage.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"strings"
 
 	"go.uber.org/zap"
 )
@@ -78,6 +79,7 @@ func (imagestorage *OnDiskLocalStorage) PutCategoryImage(id string, filename str
 
 func (imagestorage *OnDiskLocalStorage) DeleteItemImage(id string, filename string) error {
 	imagestorage.logger.Debug("Enter in filestorage DeleteItemImage()")
+	imagestorage.logger.Debug(fmt.Sprintf("name of deleting image: %s", filename))
 	err := os.Remove(imagestorage.path + "items/" + id + "/" + filename)
 	if err != nil {
 		imagestorage.logger.Debug(fmt.Sprintf("error on delete file: %v", err))
@@ -88,6 +90,7 @@ func (imagestorage *OnDiskLocalStorage) DeleteItemImage(id string, filename stri
 
 func (imagestorage *OnDiskLocalStorage) DeleteCategoryImage(id string, filename string) error {
 	imagestorage.logger.Debug("Enter in filestorage DeleteCategoryImage()")
+	imagestorage.logger.Debug(fmt.Sprintf("name of deleting image: %s", filename))
 	err := os.Remove(imagestorage.path + "categories/" + id + "/" + filename)
 	if err != nil {
 		imagestorage.logger.Debug(fmt.Sprintf("error on delete file: %v", err))
@@ -98,18 +101,23 @@ func (imagestorage *OnDiskLocalStorage) DeleteCategoryImage(id string, filename 
 
 func (imagestorage *OnDiskLocalStorage) DeleteCategoryImageById(id string) error {
 	imagestorage.logger.Debug("Enter in filestorage DeleteCategoryImageById()")
-	err := filepath.Walk(imagestorage.path+"/categories/"+id, func(path string, info os.FileInfo, err error) error {
+	imagestorage.logger.Debug(fmt.Sprintf("path is: %s", imagestorage.path+"categories/"+id))
+	err := filepath.Walk(imagestorage.path+"categories/"+id, func(path string, info os.FileInfo, err error) error {
+		imagestorage.logger.Debug(fmt.Sprintf("isDir: %t, name: %s", info.IsDir(), info.Name()))
 		if !info.IsDir() {
-			err := imagestorage.DeleteCategoryImage(id, info.Name())
-			if err != nil {
-				imagestorage.logger.Error(fmt.Sprintf("error on delete category image: %v", err))
-				return err
+			if strings.Contains(info.Name(), ".jpeg") || strings.Contains(info.Name(), ".png") || strings.Contains(info.Name(), ".jpg") {
+				imagestorage.logger.Debug(fmt.Sprintf("find file, name is: %s", info.Name()))
+				err := imagestorage.DeleteCategoryImage(id, info.Name())
+				if err != nil {
+					imagestorage.logger.Error(fmt.Sprintf("error on delete category image: %v", err))
+					return err
+				}
 			}
 		}
 		return nil
 	})
 	if err != nil {
-		imagestorage.logger.Error(fmt.Sprintf("error no filepath.Walk: %v", err))
+		imagestorage.logger.Error(fmt.Sprintf("error on filepath.Walk: %v", err))
 		return err
 	}
 	imagestorage.logger.Debug(fmt.Sprintf("images of category with id %s deleted success", id))
